feat: ADR-005 WebConfig 이관 (#413)
WebConfig, CorsProperties, OpenApiConfig를 module-app에서
module-web/web/config/로 이관

변경 사항:
- WebConfig.java: MDCFilter 등록 (module-infra 의존)
- CorsProperties.java: CORS 설정 프로퍼티
- OpenApiConfig.java: Swagger/OpenAPI 설정
