+++ module-web/src/main/java/maple/expectation/web/config/WebConfig.java
@@ -1,4 +1,4 @@
-package maple.expectation.config;
+package maple.expectation.web.config;
 
 import maple.expectation.infrastructure.filter.MDCFilter;
 import org.springframework.boot.web.servlet.FilterRegistrationBean;
@@ -6,6 +6,11 @@ import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.core.Ordered;
 
+/**
+ * Web Configuration (ADR-005 이관)
+ *
+ * <p>책임: MDCFilter 등록 및 순서 설정
+ */
 @Configuration
 public class WebConfig {
 
@@ -13,8 +18,8 @@ public class WebConfig {
   public FilterRegistrationBean<MDCFilter> mdcFilterRegistration(MDCFilter mdcFilter) {
     FilterRegistrationBean<MDCFilter> registrationBean = new FilterRegistrationBean<>(mdcFilter);
 
-    // 💡 핵심: 모든 필터 중 가장 먼저 실행되도록 순서를 최우선(0순위)으로 설정합니다.
-    // 그래야 보안 필터나 API 로직에서 발생하는 모든 로그에 requestId가 찍힙니다.
+    // 핵심: 모든 필터 중 가장 먼저 실행되도록 순서를 최우선(0순위)으로 설정
+    // 그래야 보안 필터나 API 로직에서 발생하는 모든 로그에 requestId가 찍힘
     registrationBean.setOrder(Ordered.HIGHEST_PRECEDENCE);
 
     // 모든 URL 패턴에 대해 필터 적용
